+++ src/test/java/com/guillermomolina/protos/execution/ProtosCoreNativeBoundaryArchitectureTest.java
@@ -90,7 +90,7 @@ final class ProtosCoreNativeBoundaryArchitectureTest {
             Map.ofEntries(
                     Map.entry("execution/ProtosExactExecutionFacility.java", 2),
                     Map.entry("execution/ProtosTestToolCatalogAcquisitionFacility.java", 1),
-                    Map.entry("execution/ProtosTestToolFileSelectionFacility.java", 1));
+                    Map.entry("execution/ProtosTestToolFileSelectionFacility.java", 2));
 
     @Test
     void javaNativeClosureProvidersMatchTheAuditedBoundaryExactly() throws IOException {